change argument type to ssize_t

+++ cdef/mmap/mmap.h
@@ -1,4 +1,4 @@
-void *mmap(void *addr, size_t length, int prot, int flags, int fd, off_t offset);
+void *mmap(void *addr, size_t length, int prot, int flags, int fd, ssize_t offset);
 int munmap(void *addr, size_t length);
 int msync(void *addr, size_t length, int flags);
 int mprotect(void *addr, size_t len, int prot);
